# Conflicts:
#	out/production/Ex2/Ex2.class
#	out/production/Ex2/api/DirectedWeightedGraph.class
#	out/production/Ex2/api/DirectedWeightedGraphAlgorithms.class
#	out/production/Ex2/api/Edge.class
#	out/production/Ex2/api/EdgeData.class
#	out/production/Ex2/api/GeoLocation.class
#	out/production/Ex2/api/GeoPoint.class
#	out/production/Ex2/api/Graph.class
#	out/production/Ex2/api/Node.class
#	out/production/Ex2/api/NodeData.class
#	out/production/Ex2/api/jsonParser.class

+++ src/api/GeoPoint.java
@@ -1,11 +1,11 @@
 package api;
 
 public class GeoPoint implements GeoLocation {
-    private final int x;
-    private final int y;
-    private final int z;
+    private final double x;
+    private final double y;
+    private final double z;
 
-    public GeoPoint(int x, int y, int z){
+    public GeoPoint(double x, double y, double z){
         this.x = x;
         this.y = y;
         this.z = z;
